@@ -13,29 +13,30 @@
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 from __future__ import annotations
 
-from typing import Dict, Optional, Sequence, Type, Union
-
-from pydantic import BaseModel
+from typing import Optional, Sequence, Union
 
 from camel.configs.base_config import BaseConfig
+from camel.types import NotGiven
 
 
-class ChatGPTConfig(BaseConfig):
-    r"""Defines the parameters for generating chat completions using the
-    OpenAI API.
+class OpenRouterConfig(BaseConfig):
+    r"""Defines the parameters for generating chat completions using OpenAI
+    compatibility.
+
+    Reference: https://openrouter.ai/docs/api-reference/parameters
 
     Args:
         temperature (float, optional): Sampling temperature to use, between
             :obj:`0` and :obj:`2`. Higher values make the output more random,
             while lower values make it more focused and deterministic.
-            (default: :obj:`None`)
+            (default: :obj: `None`)
         top_p (float, optional): An alternative to sampling with temperature,
             called nucleus sampling, where the model considers the results of
             the tokens with top_p probability mass. So :obj:`0.1` means only
             the tokens comprising the top 10% probability mass are considered.
-            (default: :obj:`None`)
+            (default: :obj: `None`)
         n (int, optional): How many chat completion choices to generate for
-            each input message. (default: :obj:`None`)
+            each input message. (default: :obj: `None`)
         response_format (object, optional): An object specifying the format
             that the model must output. Compatible with GPT-4 Turbo and all
             GPT-3.5 Turbo models newer than gpt-3.5-turbo-1106. Setting to
@@ -51,7 +52,7 @@ class ChatGPTConfig(BaseConfig):
             max context length.
         stream (bool, optional): If True, partial message deltas will be sent
             as data-only server-sent events as they become available.
-            (default: :obj:`None`)
+            (default: :obj: `None`)
         stop (str or list, optional): Up to :obj:`4` sequences where the API
             will stop generating further tokens. (default: :obj:`None`)
         max_tokens (int, optional): The maximum number of tokens to generate
@@ -62,28 +63,19 @@ class ChatGPTConfig(BaseConfig):
             :obj:`2.0`. Positive values penalize new tokens based on whether
             they appear in the text so far, increasing the model's likelihood
             to talk about new topics. See more information about frequency and
-            presence penalties. (default: :obj:`None`)
+            presence penalties. (default: :obj: `None`)
         frequency_penalty (float, optional): Number between :obj:`-2.0` and
             :obj:`2.0`. Positive values penalize new tokens based on their
             existing frequency in the text so far, decreasing the model's
             likelihood to repeat the same line verbatim. See more information
-            about frequency and presence penalties. (default: :obj:`None`)
-        logit_bias (dict, optional): Modify the likelihood of specified tokens
-            appearing in the completion. Accepts a json object that maps tokens
-            (specified by their token ID in the tokenizer) to an associated
-            bias value from :obj:`-100` to :obj:`100`. Mathematically, the bias
-            is added to the logits generated by the model prior to sampling.
-            The exact effect will vary per model, but values between:obj:` -1`
-            and :obj:`1` should decrease or increase likelihood of selection;
-            values like :obj:`-100` or :obj:`100` should result in a ban or
-            exclusive selection of the relevant token. (default: :obj:`None`)
+            about frequency and presence penalties. (default: :obj: `None`)
         user (str, optional): A unique identifier representing your end-user,
             which can help OpenAI to monitor and detect abuse.
-            (default: :obj:`""`)
+            (default: :obj: `None`)
         tools (list[FunctionTool], optional): A list of tools the model may
             call. Currently, only functions are supported as a tool. Use this
             to provide a list of functions the model may generate JSON inputs
-            for. A max of 128 functions are supported.
+            for. A max of 128 functions are supported. (default: :obj: `None`)
         tool_choice (Union[dict[str, str], str], optional): Controls which (if
             any) tool is called by the model. :obj:`"none"` means the model
             will not call any tool and instead generates a message.
@@ -93,33 +85,22 @@ class ChatGPTConfig(BaseConfig):
             via {"type": "function", "function": {"name": "my_function"}}
             forces the model to call that tool. :obj:`"none"` is the default
             when no tools are present. :obj:`"auto"` is the default if tools
-            are present.
-        reasoning_effort(str, optional): A parameter specifying the level of
-            reasoning used by certain model types. Valid values are :obj:
-            `"low"`, :obj:`"medium"`, or :obj:`"high"`. If set, it is only
-            applied to the model types that support it (e.g., :obj:`o1`,
-            :obj:`o1mini`, :obj:`o1preview`, :obj:`o3mini`). If not provided
-            or if the model type does not support it, this parameter is
-            ignored. (default: :obj:`None`)
-        parallel_tool_calls (bool, optional): A parameter specifying whether
-            the model should call tools in parallel or not. (default:
-            :obj:`None`)
+            are present. (default: :obj: `None`)
     """
 
     temperature: Optional[float] = None
     top_p: Optional[float] = None
     n: Optional[int] = None
     stream: Optional[bool] = None
-    stop: Optional[Union[str, Sequence[str]]] = None
-    max_tokens: Optional[int] = None
+    stop: Optional[Union[str, Sequence[str], NotGiven]] = None
+    max_tokens: Optional[Union[int, NotGiven]] = None
     presence_penalty: Optional[float] = None
-    response_format: Optional[Union[Type[BaseModel], Dict]] = None
+    response_format: Optional[Union[dict, NotGiven]] = None
     frequency_penalty: Optional[float] = None
-    logit_bias: Optional[Dict] = None
     user: Optional[str] = None
-    tool_choice: Optional[Union[Dict[str, str], str]] = None
-    reasoning_effort: Optional[str] = None
-    parallel_tool_calls: Optional[bool] = None
+    tool_choice: Optional[Union[dict[str, str], str]] = None
 
 
-OPENAI_API_PARAMS = {param for param in ChatGPTConfig.model_fields.keys()}
+OPENROUTER_API_PARAMS = {
+    param for param in OpenRouterConfig.model_fields.keys()
+}
