@@ -13,7 +13,7 @@
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 import asyncio
 
-from camel.verifiers import PythonVerifier, VerifierInput
+from camel.verifiers import PythonVerifier
 
 verifier = PythonVerifier(required_packages=["numpy"])
 asyncio.run(verifier.setup())
@@ -29,12 +29,14 @@ print(result)
 
 # Since the output of the above numpy code evaluates to 32,
 # we expect the verification outcome to be a success.
-response = VerifierInput(llm_response=numpy_test_code, ground_truth="32")
-result = asyncio.run(verifier.verify(response))
+result = asyncio.run(
+    verifier.verify(solution=numpy_test_code, ground_truth="32")
+)
 print(f"Result: {result.status}")
 
-response = VerifierInput(llm_response=numpy_test_code, ground_truth="40")
-result = asyncio.run(verifier.verify(response))
+result = asyncio.run(
+    verifier.verify(solution=numpy_test_code, ground_truth="40")
+)
 
 # Now we expect the VerificationOutcome to be a failure,
 # because the answer is wrong.
