@@ -17,7 +17,6 @@ import pytest
 
 from camel.verifiers.models import (
     VerificationOutcome,
-    VerifierInput,
 )
 from camel.verifiers.python_verifier import PythonVerifier
 
@@ -48,7 +47,7 @@ async def test_python_verifier_execution_success(python_verifier):
 
     script = 'print("Hello, World!")'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.SUCCESS
@@ -64,7 +63,7 @@ async def test_python_verifier_execution_failure(python_verifier):
 
     script = 'print("Hello"'  # Syntax error
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.ERROR
@@ -81,7 +80,7 @@ async def test_python_verifier_execution_timeout(python_verifier):
 
     script = 'import time; time.sleep(5)'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.TIMEOUT
@@ -97,7 +96,7 @@ async def test_python_verifier_output_mismatch(python_verifier):
 
     script = 'print("Wrong output")'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script, ground_truth="Expected output")
+        solution=script, ground_truth="Expected output"
     )
 
     assert result.status == VerificationOutcome.FAILURE
@@ -113,7 +112,7 @@ async def test_python_verifier_correct_output_matching(python_verifier):
 
     script = 'print("Expected output")'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script, ground_truth="Expected output")
+        solution=script, ground_truth="Expected output"
     )
 
     assert result.status == VerificationOutcome.SUCCESS
@@ -128,7 +127,7 @@ async def test_python_verifier_no_venv_error():
     verifier = PythonVerifier()
     script = 'print("Hello")'
     result = await verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.ERROR
@@ -142,7 +141,7 @@ async def test_python_verifier_with_numpy(python_verifier):
 
     script = 'import numpy as np; print(np.array([1, 2, 3]))'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.SUCCESS
